// MobileAPP/src/utilis/types/Registration.ts
export interface valueType {
  motherTongue: string;
  maritalStatus: string;
  religion: string;
  caste: string;
  interCommunity: any;
  country: any;
  city: any;
  state: any;
  Children: any;
  numberOfChildren: any;
  height: string;
  physicalStatus: string;
  Weight: string;
  bodyType: string;
  addressLine: string;

  ChildrenError: any;
  numberOfChildrenError: any;
  motherTongueError: any;
  maritalStatusError: string;
  religionError: string;
  casteError: string;
  interCommunityError: string;
  countryError: string;
  cityError: string;
  stateError: string;
  heightError: string;
  physicalStatusError: string;
  WeightError: string;
  bodyTypeError: string;
  addressLineError: string;
}

export const formValue = {
  motherTongue: '',
  maritalStatus: '',
  religion: '',
  caste: '',
  interCommunity: '',
  country: '',
  city: '',
  state: '',
  Children: '',
  numberOfChildren: '',
  height: '',
  physicalStatus: '',
  Weight: '',
  bodyType: '',
  addressLine: '',

  ChildrenError: '',
  numberOfChildrenError: '',
  motherTongueError: '',
  maritalStatusError: '',
  religionError: '',
  casteError: '',
  interCommunityError: '',
  countryError: '',
  cityError: '',
  stateError: '',
  heightError: '',
  physicalStatusError: '',
  WeightError: '',
  bodyTypeError: '',
  addressLineError: '',
};